+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### vNext
 
-* Make WrapQuery work for non-root fields <br />
+* Make `WrapQuery` work for non-root fields <br />
   [@mdlavin](https://github.com/mdlavin) in
   [#1007](https://github.com/apollographql/graphql-tools/pull/1008)