+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+### vNext
+
+* Make `WrapQuery` work for non-root fields <br />
+  [@mdlavin](https://github.com/mdlavin) in
+  [#1007](https://github.com/apollographql/graphql-tools/pull/1008)
+
 ### 4.0.3
 
 * Replaced broken link in docs homepage with Launchpad example <br />
@@ -17,7 +23,7 @@
 * Changes to `extractExtensionDefinitions` to support `graphql-js` union and enum extensions.  <br/>
   [@jansuchy](https://github.com/jansuchy) in [#951](https://github.com/apollographql/graphql-tools/pull/951)
 * Add docs for `mockServer` (closes [#951](https://github.com/apollographql/graphql-tools/issues/94))<br/>
-  [@mfix22](https://github.com/mfix22) in [PR #982](https://github.com/apollographql/graphql-tools/pull/982)  
+  [@mfix22](https://github.com/mfix22) in [PR #982](https://github.com/apollographql/graphql-tools/pull/982)
 * Fix regression where custom scalars were incorrectly replaced while recreating schema with `visitSchema`.  <br/>
   [@tgriesser](https://github.com/tgriesser) in [#985](https://github.com/apollographql/graphql-tools/pull/985)